chore: rename ExpectedAnswerCardView
